requirements.txt
Django==6.0.3
requests==2.32.5
psycopg2-binary==2.9.11
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1